Fix UnboundLocalError exception
Use a different variable name for the event loop in the inner function,
so that it can be assigned to without affecting the outer scope.

Using the same variable name is not possible,
because the variable is assigned to,
which makes it a local variable.
Therefore, it must be assigned a value *before* it can be used,
which makes it impossible to at first check whether it is None.

This principle is also the reason why
the original code raised UnboundLocalError exception.

It is possible to make the loop variable nonlocal,
but then the assignments would be visible to the outer scope.
The consequence would be that when loop is None,
an event loop assigned to the loop variable
in the first excepthook invocation
would be used in all the subsequent excepthook invocations.

I do not know whether it would be a good idea,
because perhaps asyncio.get_event_loop may return different event loops
during different excepthook invocations.
Therefore, I consider it safer
to always get the event loop
directly in the excepthook.